+++ NEWS.md
@@ -14,6 +14,12 @@ for human readability.
 #### Deprecated
 
 #### Removed
+
+- The shallow-water equation types `ShallowWaterEquations1D`, `ShallowWaterEquations2D`, and 
+  `ShallowWaterEquationsQuasi1D` have been removed from Trixi.jl and are now available via 
+  [TrixiShallowWater.jl](https://github.com/trixi-framework/TrixiShallowWater.jl/). 
+  This also affects the related functions `hydrostatic_reconstruction_audusse_etal`, 
+  `flux_nonconservative_audusse_etal`, and `FluxHydrostaticReconstruction`. ([#2379])
 - The additional `ìnitial_cache` entries in the caches of `SemidiscretizationHyperbolic`
   and `SemidiscretizationHyperbolicParabolic`, and the corresponding keyword arguments of
   their constructors have been removed. ([#2399])
@@ -22,6 +28,10 @@ for human readability.
 
 #### Added
 
+- Added symmetry plane/reflective wall velocity+stress boundary conditions for the compressible Navier-Stokes equations in 2D and 3D. 
+  Currently available only for the `P4estMesh` mesh type, `GradientVariablesPrimitive`, and `Adiabatic` heat boundary condition ([#2416]).
+- Added `LaplaceDiffusionEntropyVariables1D`, `LaplaceDiffusionEntropyVariables2D`, and `LaplaceDiffusionEntropyVariables3D`. These add scalar diffusion to each
+  equation of a system, but apply diffusion in terms of the entropy variables, which symmetrizes the viscous formulation and ensures semi-discrete entropy dissipation ([#2406]).
 - Added the three-dimensional multi-ion magneto-hydrodynamics (MHD) equations with a
   generalized Lagrange multipliers (GLM) divergence cleaning technique ([#2215]).
 - New time integrator `PairedExplicitRK4`, implementing the fourth-order